+++ packages/mg-components/src/components/atoms/mg-badge/doc/mg-badge.stories.tsx
@@ -1,6 +1,7 @@
 import { h } from '@stencil/core';
-import { filterArgs } from '../../../../../.storybook/utils';
+import { filterArgs } from '@mgdis/stencil-helpers';
 import { variants } from '../mg-badge.conf';
+import type { MgBadge as MgBadgeType } from '../mg-badge';
 
 export default {
   component: 'mg-badge',
@@ -17,8 +18,7 @@ export default {
  * @param args - component arguments
  * @returns HTMLElement
  */
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-const Template = (args: any): HTMLElement => <mg-badge {...filterArgs(args, { variant: variants[0] })}></mg-badge>;
+const Template = (args: MgBadgeType): HTMLElement => <mg-badge {...filterArgs(args, { variant: variants[0] })}></mg-badge>;
 
 export const MgBadge = {
   render: Template,