+++ packages/mg-components/src/components/molecules/menu/mg-menu-item/mg-menu-item.tsx
@@ -1,5 +1,5 @@
 import { Component, h, Prop, State, Host, Watch, Element, Event, EventEmitter } from '@stencil/core';
-import { ClassList, createID, isValidString } from '../../../../utils/components.utils';
+import { ClassList, createID, isValidString } from '@mgdis/stencil-helpers';
 import { initLocales } from '../../../../locales';
 import { Direction } from '../mg-menu/mg-menu.conf';
 import { Status } from './mg-menu-item.conf';