+++ packages/mg-components/src/components/molecules/mg-panel/test/mg-panel.e2e.ts
@@ -1,6 +1,6 @@
 import { setPageContent, expect, describe, testEach, updateScreenshotClass, PageType, test } from '../../../../utils/playwright.e2e.test.utils';
-import { createID } from '../../../../utils/components.utils';
-import { renderAttributes, renderProperties } from '../../../../utils/e2e.test.utils';
+import { createID } from '@mgdis/stencil-helpers';
+import { renderAttributes, renderProperties } from '@mgdis/playwright-helpers';
 import { type MgPanel } from '../mg-panel';
 
 type ArgsType = Partial<MgPanel> & { slot?: SlotType };